+++ Sources/_MapboxNavigationTestHelpers/Core/AlternativeRouteMock.swift
@@ -5,57 +5,57 @@ import MapboxNavigationNative_Private
 import Turf
 
 extension AlternativeRoute {
-    public static func mock(
-        mainRoute: Route = .mock(),
-        alternativeRoute: Route = .mock(shape: .mock(delta: (0.01, -0.01))),
-        nativeRouteAlternative: RouteAlternative? = nil,
-        requestOptions: ResponseOptions? = nil,
-        routeIndex: Int = 0
-    ) -> Self {
-        let nativeAlternative = nativeRouteAlternative ?? .mock(
-            mainRoute: mainRoute,
-            alternativeRoute: alternativeRoute,
-            routeIndex: routeIndex
-        )
-        let urlString = nativeAlternative.route.getRequestUri()
-        let requestOptions = requestOptions ?? .mock(routeOptions: .mock(string: urlString))
-        return self.init(
-            mainRoute: mainRoute,
-            alternativeRoute: alternativeRoute,
-            nativeRouteAlternative: nativeAlternative,
-            requestOptions: requestOptions
-        )!
-    }
+//    public static func mock(
+//        mainRoute: Route = .mock(),
+//        alternativeRoute: Route = .mock(shape: .mock(delta: (0.01, -0.01))),
+//        nativeRouteAlternative: RouteAlternative? = nil,
+//        requestOptions: ResponseOptions? = nil,
+//        routeIndex: Int = 0
+//    ) -> Self {
+//        let nativeAlternative = nativeRouteAlternative ?? .mock(
+//            mainRoute: mainRoute,
+//            alternativeRoute: alternativeRoute,
+//            routeIndex: routeIndex
+//        )
+//        let urlString = nativeAlternative.route.getRequestUri()
+//        let requestOptions = requestOptions ?? .mock(routeOptions: .mock(string: urlString))
+//        return self.init(
+//            mainRoute: mainRoute,
+//            alternativeRoute: alternativeRoute,
+//            nativeRouteAlternative: nativeAlternative,
+//            requestOptions: requestOptions
+//        )!
+//    }
 }
 
 extension RouteAlternative {
-    public static func mock(mainRoute: Route, alternativeRoute: Route, routeIndex: Int = 0) -> Self {
-        .mock(route: RouteInterfaceMock(
-            route: mainRoute,
-            alternativeRoute: alternativeRoute,
-            routeIndex: routeIndex
-        ))
-    }
-
-    public static func mock(
-        id: UInt32 = 0,
-        route: RouteInterface = RouteInterfaceMock(),
-        mainRouteFork: RouteIntersection = .mock(),
-        alternativeRouteFork: RouteIntersection = .mock(),
-        infoFromFork: AlternativeRouteInfo = .mock(),
-        infoFromStart: AlternativeRouteInfo = .mock(),
-        isNew: Bool = true
-    ) -> Self {
-        self.init(
-            id: id,
-            route: route,
-            mainRouteFork: mainRouteFork,
-            alternativeRouteFork: alternativeRouteFork,
-            infoFromFork: infoFromFork,
-            infoFromStart: infoFromStart,
-            isNew: isNew
-        )
-    }
+//    public static func mock(mainRoute: Route, alternativeRoute: Route, routeIndex: Int = 0) -> Self {
+//        .mock(route: RouteInterfaceMock(
+//            route: mainRoute,
+//            alternativeRoute: alternativeRoute,
+//            routeIndex: routeIndex
+//        ))
+//    }
+//
+//    public static func mock(
+//        id: UInt32 = 0,
+//        route: RouteInterface = RouteInterfaceMock(),
+//        mainRouteFork: RouteIntersection = .mock(),
+//        alternativeRouteFork: RouteIntersection = .mock(),
+//        infoFromFork: AlternativeRouteInfo = .mock(),
+//        infoFromStart: AlternativeRouteInfo = .mock(),
+//        isNew: Bool = true
+//    ) -> Self {
+//        self.init(
+//            id: id,
+//            route: route,
+//            mainRouteFork: mainRouteFork,
+//            alternativeRouteFork: alternativeRouteFork,
+//            infoFromFork: infoFromFork,
+//            infoFromStart: infoFromStart,
+//            isNew: isNew
+//        )
+//    }
 }
 
 extension RouteIntersection {